+++ spec/transpec/syntax/should_spec.rb
@@ -226,6 +226,40 @@ module Transpec
             record.converted_syntax.should == 'expect(obj).to'
           end
 
+          context 'and #expect is available in the context by including RSpec::Matchers' do
+            let(:source) do
+              <<-END
+                describe 'example' do
+                  class TestRunner
+                    include RSpec::Matchers
+
+                    def run
+                      1.should == 1
+                    end
+                  end
+
+                  it 'is 1' do
+                    TestRunner.new.run
+                  end
+                end
+              END
+            end
+
+            context 'with runtime information' do
+              include_context 'dynamic analysis objects'
+
+              it 'does not raise InvalidContextError' do
+                -> { should_object.expectize! }.should_not raise_error
+              end
+            end
+
+            context 'without runtime information' do
+              it 'raises InvalidContextError' do
+                -> { should_object.expectize! }.should raise_error(InvalidContextError)
+              end
+            end
+          end
+
           context 'and #expect is not available in the context' do
             context 'and the context is determinable statically' do
               let(:source) do